+++ CHANGELOG.md
@@ -143,8 +143,6 @@
 - WMS-Time:
     - The attributes `dimensionRange` and `dimensionRegex` can now be used to filter the values ​​that should be displayed in the time slider.
     - The attibute `staticDimensions` can now be configured and sent to the service when the WMS time layer is loaded.
-=======
-- Issue \#1524: draw_old: Added documentation for drawSquareSettings.
 - WMS-Time: The attributes `dimensionRange` and `dimensionRegex` can now be used to filter the values ​​that should be displayed in the time slider.
 - CopyrightConstraints: Added Parameter `useLayerCswUrl` to use the CSW interface specified in the layers.
 - The following packages have been added:
@@ -163,6 +161,7 @@
     - vite-plugin-node-polyfills: ^0.24.0
     - vite-plugin-zip-pack: ^1.2.4
     - vitest: ^4.0.16
+- Filter: Adds parameter shouldSubjectDataSelectedInTree: Filters are only displayed when the respective layer is selected in the subject tree
 
 ### Changed
 - Issue \#1500: WfsSearch: Added multi-select functionality for search results with polygon markers on map and zoom to combined extent. Improved geometry extraction and row selection logic to prevent duplicate entries.